Reenabled history, empty refresh yields logo, added welcome message and other minor changes.
There is a new save button to add the current equation to history, once
it is pressed, the new history is echoed to the terminal.

If the equation box is empty or contains the welcome message when the
refresh button is pressed, instead of this rendered to the imageView,
the SemTeX logo reappears.

Finally, I played with some of the widget sizes and made some small
clean ups to the code as well as adding .hist to .gitignore.