Added an "About SemTex" option to the help menu.
Help menu has replaced About menu, Help contains "About SemTeX" which has the logo, tag line, says it was created by me and has links to my email and semtex github page.

I also fixed the position of the main window when it opens because it kept opening at the very bottom of the screen when called from pycharm.